refactor: update validation imports and introduce new validation schema for CustomPages
